+++ NEWS.md
@@ -1,3 +1,5 @@
+# legendry (development version)
+
 # legendry 0.1.0
 
 First release.